4-MAR-24-FARIZD-RUSNI

+++ js/script.js
@@ -2,7 +2,7 @@ function Calculate() {
   var height = document.getElementById("h-input").value;
   var weight = document.getElementById("w-input").value;
 
-  var result = parseFloat(weight) / (parseFloat(height) / 100) ** 2;
+  var result = parseInt(weight) / (parseInt(height) / 100) ** 2;
 
   if (!isNaN(result)) {
     document.getElementById("bmi-output").innerHTML = result;
